refactor: standardize JsonLd imports to kebab-case, remove duplicate component
