+++ app/layout.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from "next"
-import JsonLd from "@/components/JsonLd"
+import { JsonLd } from "@/components/json-ld"
 import { Inter } from 'next/font/google'
 import Script from 'next/script'
 import { Analytics } from "@vercel/analytics/next"
