--- components/JsonLd.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-type Props = { data: Record<string, unknown> };
-export default function JsonLd({ data }: Props) {
-  return (
-    <script
-      type="application/ld+json"
-      dangerouslySetInnerHTML={{ __html: JSON.stringify(data) }}
-    />
-  );
-}
